fix: ensure probability service initialization
